+++ yafuse-fork/libext4/inode.c
@@ -110,7 +110,7 @@ int32_t ext4_fill_inode(const struct ext4_super_block *sb, const struct ext4_gro
     offset = inode_tbl + (inode_num * sb->s_inode_size);
   }
 
-  ret = io_fseek(offset);
+  ret = io_fseek((long)offset);
   if (ret != 0) {
     return -1;
   }
